# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baze API",
    "description": "Collaborative Bicycle Navigation REST & Realtime API",
    "contact": {
      "name": "Gabriel Piñones"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/anonymous": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "anonymous_auth_handler",
        "responses": {
          "201": {
            "description": "Anonymous account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/geocoding/search": {
      "get": {
        "tags": [
          "geocoding"
        ],
        "operationId": "geocoding_handler",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Geocoding suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeocodingItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Address search is not available yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Geocoding engine failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hazards": {
      "get": {
        "tags": [
          "hazards"
        ],
        "operationId": "list_hazards_handler",
        "parameters": [
          {
            "name": "min_lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "min_lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "max_lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "max_lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active hazards within bounding box",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hazard"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid bbox query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hazards"
        ],
        "operationId": "create_hazard_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHazardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Hazard report created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hazard"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "The hazard store is full",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/hazards/{id}/vote": {
      "post": {
        "tags": [
          "hazards"
        ],
        "operationId": "vote_hazard_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Hazard ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HazardVoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Vote registered and hazard updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hazard"
                }
              }
            }
          },
          "400": {
            "description": "Invalid vote payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Hazard not found or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/realtime/sse": {
      "get": {
        "tags": [
          "realtime"
        ],
        "operationId": "realtime_sse_handler",
        "parameters": [
          {
            "name": "min_lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "min_lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "max_lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "max_lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events stream of hazards within bounding box",
            "content": {
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "Invalid bbox query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit or concurrent connection limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/routing/route": {
      "post": {
        "tags": [
          "routing"
        ],
        "operationId": "route_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Calculated bicycle route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid coordinates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Routing is not available yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Routing engine failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "type": "object",
        "description": "Response to creating an anonymous account.",
        "required": [
          "account_id",
          "token",
          "created_at",
          "expires_at"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "The token stops working at this instant; create a new account afterwards."
          },
          "token": {
            "type": "string",
            "description": "Bearer token: `baze_v2.<payload>.<mac>`. Send it as `Authorization: Bearer <token>`."
          }
        }
      },
...
      "CreateHazardRequest": {
        "type": "object",
        "required": [
          "category",
          "location"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/HazardCategory"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "$ref": "#/components/schemas/GeoJsonPoint"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable reason. For 5xx responses it is deliberately generic."
          },
          "error_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Present on 5xx responses: quote it when reporting a problem, the details are in the server log under this id."
          }
        }
      },
...
      "GeoJsonPoint": {
        "type": "object",
        "required": [
          "type",
          "coordinates"
        ],
        "properties": {
          "coordinates": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            },
            "description": "[longitude, latitude]"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "GeocodingItem": {
        "type": "object",
        "required": [
          "name",
          "location"
        ],
        "properties": {
          "city": {
            "type": [
              "string",
              "null"
            ]
          },
          "country": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "$ref": "#/components/schemas/GeoJsonPoint"
          },
          "name": {
            "type": "string"
          },
          "street": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Hazard": {
        "type": "object",
        "required": [
          "id",
          "category",
          "hazard_type",
          "status",
          "upvotes",
          "downvotes",
          "location",
          "created_at",
          "expires_at"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/HazardCategory"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "downvotes": {
            "type": "integer",
            "format": "int32"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "hazard_type": {
            "$ref": "#/components/schemas/HazardType"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "location": {
            "$ref": "#/components/schemas/GeoJsonPoint"
          },
          "status": {
            "$ref": "#/components/schemas/HazardStatus"
          },
          "upvotes": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "HazardCategory": {
        "type": "string",
        "enum": [
          "glass",
          "pothole",
          "debris",
          "road_closed",
          "construction",
          "flood"
        ]
      },
...
      "HazardType": {
        "type": "string",
        "enum": [
          "warning",
          "blocking"
        ]
      },
      "HazardVoteRequest": {
        "type": "object",
        "required": [
          "vote"
        ],
        "properties": {
          "vote": {
            "$ref": "#/components/schemas/Vote",
            "description": "1 for upvote, -1 for downvote"
          }
        }
      },
...
      "RouteRequest": {
        "type": "object",
        "required": [
          "origin",
          "destination"
        ],
        "properties": {
          "destination": {
            "$ref": "#/components/schemas/GeoJsonPoint"
          },
          "origin": {
            "$ref": "#/components/schemas/GeoJsonPoint"
          }
        }
      },
      "RouteResponse": {
        "type": "object",
        "required": [
          "distance_meters",
          "duration_seconds",
          "ascent_meters",
          "descent_meters",
          "geometry",
          "maneuvers",
          "nearby_hazards"
        ],
        "properties": {
          "ascent_meters": {
            "type": "number",
            "format": "double"
          },
          "descent_meters": {
            "type": "number",
            "format": "double"
          },
          "distance_meters": {
            "type": "number",
            "format": "double"
          },
          "duration_seconds": {
            "type": "number",
            "format": "double"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeoJsonLineString"
          },
          "maneuvers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteManeuver"
            }
          },
          "nearby_hazards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Hazard"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Anonymous bearer token returned by POST /api/v1/auth/anonymous, in the form baze_v2.<payload>.<mac>. It expires (see expires_at)."
      }
    }
  },
...
}